+++ src/settings.js
@@ -1626,6 +1626,9 @@ var EMBIND = 0;
 // Whether the main() function reads the argc/argv parameters.
 var MAIN_READS_PARAMS = 1;
 
+// The computed location of the pointer to the sbrk position.
+var DYNAMICTOP_PTR = -1;
+
 // Legacy settings that have been removed or renamed.
 // For renamed settings the format is:
 // [OLD_NAME, NEW_NAME]